Rework the vim importer to support file navigation as well
Instead of using tags for imports the importer now checks the g:as_locations variable to find
ActionScript classes and top-level functions. It can finagle the information it needs out of source
trees, asdocs, and swc files. For example, I have

let g:as_locations = [
\ "~/Documents/flex3_langref/all-classes.html",
\ "~/Documents/flex3_langref/flash/utils/package-detail.html",
\ "~/dev/assemblage/aspirin/src/main/as",
\ "~/dev/assemblage/flashbang/src/main/as",
\ "~/dev/assemblage/narya/src/main/as",
\ "~/dev/assemblage/nenya/src/main/as",
\ "~/dev/assemblage/vilya/src/main/as",
\ "~/dev/assemblage/orth/src/main/as",
\ "~/dev/as3-signals/src",
\ "~/dev/src/AsWing/src",
\ "~/dev/who/src/main/as"
\ ]

in my .vimrc. That lets me navigate and import all of our source and the builtin flash stuff.

The locations are scanned on the first navigation or import, after which they're cached in memory.
If an import or navigation isn't found or g:as_locations changes, the cache is flushed. On my
system, looking up the couple thousand classes in those directories takes about a tenth of a second,
and a cached lookup is instantaneous.

It's also now compatible with pathogen(http://www.vim.org/scripts/script.php?script_id=2332), which
is the easiest way to install the plugin: softlink from the vim directory here to
~/.vim/bundles/aspirin. After that, <Leader>i attempts to import the class under the cursor and
<Leader>j jumps to the class under the cursor.



git-svn-id: http://ooo-aspirin.googlecode.com/svn/trunk@161 a7ef2e1c-f4ad-2afa-6fbf-2ca91001eb2f