Automatically add missing actionscript imports when saving
The class-finding re assumes classes start with a capital, so this won't import top-level functions
like getTimer. Besides that, it does a pretty good job of finding classes.

It doesn't rescan the classes to keep it quick in the case of false-positives. If a class has been
added since the last scan, it'll have to be imported with AspirinImport.



git-svn-id: http://ooo-aspirin.googlecode.com/svn/trunk@170 a7ef2e1c-f4ad-2afa-6fbf-2ca91001eb2f